♻️ [Refactoring]: #213 RG ディレクトリを rgb/ に統合
macOS case-insensitive FS で RG/ と rg/ が衝突するため、gray/ と同じ
パターンで color/rgb/ に統合する。本コミットはリネームと import パスの
書き換えのみで、振る舞いの変更はない。

- RG/index.ts → rgb/stroke.ts
- RG/__tests__/RG.{basic,error,integration}.test.ts → rgb/__tests__/stroke.{basic,error,integration}.test.ts
- stroke 系 3 テストの import を "../index" → "../stroke" に更新

Refs #213
