+++ packages/core/src/content-stream/operators/color/rgb/__tests__/stroke.error.test.ts
@@ -3,7 +3,7 @@ import type { PdfObject } from "../../../../../pdf/types/pdf-types/index";
 import { GraphicsStateStack } from "../../../../graphics-state/index";
 import { OperandStack } from "../../../../operand-stack/index";
 import type { OperatorHandlerContext } from "../../../../operator-registry/index";
-import { RGHandler } from "../index";
+import { RGHandler } from "../stroke";
 
 const buildContext = (operands: PdfObject[]): OperatorHandlerContext => {
   const operandStack = OperandStack.create();
