+++ packages/core/src/content-stream/operators/color/rgb/stroke.ts
